Split out modules
- Split shader stuff and a couple of small utils out of the main file in preparation for fixing Vulkan compatibility.
